MM-9635: expunge global mm_config from error page
